Add files via upload
Disclaimer

This file is provided strictly for personal, non-commercial use. It was derived from content extracted from Star Citizen, which remains the intellectual property of Cloud Imperium Games (CIG). By downloading or using this file, you agree to adhere to the following terms:

No Commercial Use: This file and any 3D prints created from it are intended solely for private, non-commercial purposes and may not be sold, distributed, or used commercially in any form.

Copyright Notice: All original assets, trademarks, and designs are the property of Cloud Imperium Games (CIG). This file is not an official product of CIG and is neither affiliated with, endorsed by, nor approved by them.

User Responsibility: Users are responsible for ensuring that their use of this file complies with all applicable laws and the Star Citizen End User License Agreement (EULA).

If you are a representative of CIG and would like this file removed, please contact us, and we will take prompt action to comply with your request.